requirements.txt
groq
python-dotenv
requests
beautifulsoup4
ipython
jupyter
streamlit
